==> scripts/2-analysis/1-SOM/s01-02-som1-performance-combine-syntheticspace.R <==
library(readr)
library(tibble)
library(purrr)

# combine the performance files
df = list.files("./som_performance/", pattern = ".rds", full.names = T) |> 
  map_dfr(readRDS) 

# write out
write_rds(x = df, file = "./som_performance_combined.rds")
